refatorar código

+++ src/Pessoa.java
@@ -1,6 +1,6 @@
 public class Pessoa {
     public String nome;
-    public int idade;
+    public short idade;
     public double altura;
     public double peso;
 }
